"Refactored Signup Feature to Use React Hook Form and Updated Dependencies"
- Added `react-hook-form` to project dependencies in both `package.json` and `package-lock.json`.
- Removed the `SignupButton` component and its related logic from `src/components/Signup.jsx`, as it was no longer needed after refactoring.
- Updated `src/pages/SignupPage.jsx` to utilize `react-hook-form` for form handling instead of manual state management with Recoil atoms.
- Implemented form submission logic within `SignupPage.jsx` using `react-hook-form`'s `useForm` hook, including registration of input fields and error handling.
- Removed Recoil state hooks (`useRecoilState`, `useSetRecoilState`) previously used for managing username and password states in favor of local form state managed by `react-hook-form`.
- Integrated Axios instance for API calls directly within the `onSubmit` function of the form, replacing the previous approach used in the removed `SignupButton` component.
- Maintained navigation functionality using `useNavigate` from `react-router-dom` to redirect users upon successful signup.
