+++ package.json
@@ -15,6 +15,7 @@
     "jwt-js-decode": "^1.9.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-hook-form": "^7.52.1",
     "react-router-dom": "^6.24.1",
     "recoil": "^0.7.7"
   },
